legacy-python/requirements.txt
PyQt5>=5.15.0
psutil>=5.9.0
wmi>=1.5.1
pycryptodome>=3.18.0
pyinstaller>=5.0
